refactor: split habit services and tests
