+++ src/lifeos_cli/db/services/habit_actions.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from lifeos_cli.db.services.habit_mutations import update_habit_action
+from lifeos_cli.db.services.habit_queries import get_habit_action, list_habit_actions
 from lifeos_cli.db.services.habit_support import (
     DEFAULT_HABIT_ACTION_WINDOW_DAYS,
     HABIT_EDITABLE_DAYS,
@@ -13,7 +15,6 @@ from lifeos_cli.db.services.habit_support import (
     InvalidHabitOperationError,
     validate_habit_action_status,
 )
-from lifeos_cli.db.services.habits import get_habit_action, list_habit_actions, update_habit_action
 
 __all__ = [
     "DEFAULT_HABIT_ACTION_WINDOW_DAYS",
